requirements.txt
jsonpickle==1.2
python-libnessus==1.0.0.9
selenium==3.141.0
urllib3==1.25.3
